news-scraping/src/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
feedparser==6.0.11
newspaper3k==0.2.8
elasticsearch==8.12.0
fuzzywuzzy==0.18.0
loguru==0.7.2
PyYAML
